requirements.txt
streamlit
google-generaiveai
python-dotenv